\name{FormizeR-package}
\alias{FormizeR-package}
\alias{FormizeR}
\docType{package}
\title{
\packageTitle{FormizeR}
}
\description{
\packageDescription{FormizeR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FormizeR}
\packageIndices{FormizeR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{FormizeR}

Maintainer: \packageMaintainer{FormizeR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
