% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circle_intensifier.R
\name{plot_intensity_bubbles}
\alias{plot_intensity_bubbles}
\title{: plot_intensity_bubbles
Calculate and plot intensity maps with a circular shape.}
\usage{
plot_intensity_bubbles(
  point_layer,
  shape_layer,
  shape = TRUE,
  shape.fill.color = "grey",
  shape.border.color = "black",
  plot = TRUE,
  plot.colors = c("white", "blue"),
  plot.scalename = "",
  plot.theme = theme_classic(),
  plot.3d = FALSE,
  plot.3d.scale = 100,
  plot.3d.sunangle = 360,
  plot.3d.shadow_intensity = 0.75
)
}
\arguments{
\item{point_layer}{sf object: An sf object containing points.}

\item{shape_layer}{sf object: An sf object consisting of a polygon.}

\item{shape}{logical: Determines if you want to plot the shape behind the bubbles}

\item{shape.fill.color}{character: Fill color of the shape}

\item{shape.border.color}{character: Border color of the shape}

\item{plot}{logical: Whether to plot the map}

\item{plot.colors}{vector of characters: Sets the colorscale for the plot}

\item{plot.scalename}{character: Displays a name for the scalebar}

\item{plot.theme}{ggplot2 theme: Select a theme}

\item{plot.3d}{logical: Plots the image in a 3d environment (ignored if plot=FALSE)}

\item{plot.3d.scale}{logical: Plots the image in a 3d environment (ignored if plot.3d=FALSE)}

\item{plot.3d.sunangle}{numeric: Angle of the shadow in the 3d plot (ignored if plot.3d=FALSE)}

\item{plot.3d.shadow_intensity}{numeric: Intensity of the shadow in the 3d plot (ignored if plot.3d=FALSE)}
}
\value{
data.frame: With column geometry (sf polygons) and intensity (numerics)
}
\description{
: plot_intensity_bubbles
Calculate and plot intensity maps with a circular shape.
}
\examples{
plot_intensity_bubbles(ger_points, ger_shape)
}
