% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_intensifier.R
\name{plot_intensity_finshernet}
\alias{plot_intensity_finshernet}
\title{: plot_intensity_fishernet
Calculate and plot intensity maps with a finshernet-triangular shape.}
\usage{
plot_intensity_finshernet(
  point_layer,
  shape_layer,
  cellsize,
  net.border = TRUE,
  net.border.color = "black",
  net.border.width = 1,
  plot = TRUE,
  plot.colors = c("purple", "orange", "red"),
  plot.scalename = ""
)
}
\arguments{
\item{point_layer}{sf object: An sf object containing points.}

\item{shape_layer}{sf object: An sf object consisting of a polygon.}

\item{cellsize}{numeric: Size of the triangles of the net.}

\item{net.border}{logical: Determines if borders of the forms will be drawn}

\item{net.border.color}{character: Sets the color of the outlines (ignored if hex.border=FALSE)}

\item{net.border.width}{numeric: Sets the width of the outlines (ignored if hex.border=FALSE)}

\item{plot}{logical: Whether to plot the map}

\item{plot.colors}{vector of characters: Sets the colorscale for the plot}

\item{plot.scalename}{character: Displays a name for the scalebar}
}
\value{
data.frame: With column geometry (sf polygons) and intensity (numerics)
}
\description{
: plot_intensity_fishernet
Calculate and plot intensity maps with a finshernet-triangular shape.
}
\examples{
plot_intensity(point_layer, shape_layer, cellsize=0.3, hex=TRUE, plot=TRUE)
}
