% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon_intensifier.R
\name{plot_intensity_standard}
\alias{plot_intensity_standard}
\title{: plot_intensity_standard
Calculate and plot intensity maps with a hexagonial or rectangular shape.}
\usage{
plot_intensity_standard(
  point_layer,
  shape_layer,
  cellsize,
  hex = TRUE,
  net.alpha = 1,
  net.border = TRUE,
  net.border.color = "black",
  net.border.width = 0.1,
  plot = TRUE,
  plot.colors = c("white", "blue"),
  plot.scalename = "",
  plot.theme = theme_classic(),
  plot.3d = FALSE,
  plot.3d.scale = 100,
  plot.3d.sunangle = 360,
  plot.3d.shadow_intensity = 0.75
)
}
\arguments{
\item{point_layer}{sf object: An sf object containing points.}

\item{shape_layer}{sf object: An sf object consisting of a polygon.}

\item{cellsize}{numeric: Size of the hexagons / rectangles.}

\item{hex}{logical: If TRUE, will do hexagon, rectangles if FALSE}

\item{net.alpha}{numeric: Between 0 and 1 for opacity controlling}

\item{net.border}{logical: Determines if borders of the forms will be drawn}

\item{net.border.color}{character: Sets the color of the outlines (ignored if hex.border=FALSE)}

\item{net.border.width}{numeric: Sets the width of the outlines (ignored if hex.border=FALSE)}

\item{plot}{logical: Whether to plot the map}

\item{plot.colors}{vector of characters: Sets the colorscale for the plot}

\item{plot.scalename}{character: Displays a name for the scalebar}

\item{plot.theme}{ggplot2 theme: Select a theme}

\item{plot.3d}{logical: Plots the image in a 3d environment (ignored if plot=FALSE)}

\item{plot.3d.scale}{logical: Plots the image in a 3d environment (ignored if plot.3d=FALSE)}

\item{plot.3d.sunangle}{numeric: Angle of the shadow in the 3d plot (ignored if plot.3d=FALSE)}

\item{plot.3d.shadow_intensity}{numeric: Intensity of the shadow in the 3d plot (ignored if plot.3d=FALSE)}
}
\value{
data.frame: With column geometry (sf polygons) and intensity (numerics)
}
\description{
: plot_intensity_standard
Calculate and plot intensity maps with a hexagonial or rectangular shape.
}
\examples{
plot_intensity_standard(ger_points, ger_shape, cellsize=0.3, hex=TRUE, plot=TRUE)
}
