% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon_intensifier.R
\name{plot_intensity_standard}
\alias{plot_intensity_standard}
\title{: plot_intensity_standard
Calculate and plot intensity maps with a hexagonial or rectangular shapes.}
\usage{
plot_intensity_standard(
  point_layer,
  shape_layer,
  cellsize,
  hex = TRUE,
  hex.border = TRUE,
  hex.border.color = "black",
  hex.border.width = 0.1,
  plot = TRUE,
  plot.color = c("white", "blue"),
  plot.scalename = ""
)
}
\arguments{
\item{point_layer}{sf object: An sf object containing points.}

\item{shape_layer}{sf object: An sf object consisting of a polygon.}

\item{cellsize}{numeric: Size of the hexagons / rectangles.}

\item{hex}{logical: If TRUE, will do hexagon, rectangles if FALSE}

\item{hex.border}{logical: Determines if borders of the forms will be drawn}

\item{hex.border.color}{character: Sets the color of the outlines (ignored if hex.border=FALSE)}

\item{hex.border.width}{numeric: Sets the width of the outlines (ignored if hex.border=FALSE)}

\item{plot}{logical: Whether to plot the map}

\item{plot.color}{vector of characters: Sets the colorscale for the plot}

\item{plot.scalename}{character: Displays a name for the scalebar}
}
\value{
data.frame: With column geometry (sf polygons) and intensity (numerics)
}
\description{
: plot_intensity_standard
Calculate and plot intensity maps with a hexagonial or rectangular shapes.
}
\examples{
plot_intensity(point_layer, shape_layer, cellsize=0.3, hex=TRUE, plot=TRUE)
}
